refactor: Rename entity service -> entities service
Following naming convention as the plural 'entities' seems better
