+++ apps/members/src/members.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
-import { Member } from '../../../libs/entity/src/members/member.entity'
+import { Member } from '../../../libs/entities/src/members/member.entity'
 import { Repository } from 'typeorm'
 
 @Injectable()
