+++ libs/entities/src/attendances/attendance.entity.ts
