+++ libs/entities/src/common/abstract.entity.ts
