+++ libs/entities/src/entities.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common'
+import { EntitiesService } from './entities.service'
+
+@Module({
+	providers: [EntitiesService],
+	exports: [EntitiesService],
+})
+export class EntitiesModule {}
