+++ libs/entities/src/index.ts
@@ -1,0 +1,2 @@
+export * from './entities.module'
+export * from './entities.service'
