+++ libs/entities/src/meetings/meeting.entity.ts
