+++ libs/entities/src/members/member.entity.ts
