--- libs/entity/src/entity.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common'
-import { EntityService } from './entity.service'
-
-@Module({
-	providers: [EntityService],
-	exports: [EntityService],
-})
-export class EntityModule {}
