--- libs/entity/src/index.ts
@@ -1,2 +1,0 @@
-export * from './entity.module'
-export * from './entity.service'
